Initial commit
At this point in development, the bot has a command called >fentopng
that takes a chess position in FEN notation, and draws a chess board
with the pieces on it. It expects the FEN to be of a legal position,
otherwise the check detection mechanism can be akward at times.

Signed-off-by: ZweiSpeedruns <zcm727@gmail.com>